Fix test scripts.

+++ test/test_parser.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
+# Check comparable file.
+DIFF_FILE="test/test_result.csv"
+if [ ! -e $DIFF_FILE ]; then
+    echo "No diff file($DIFF_FILE)."
+    exit 1
+fi
+
 # Black Box test for cron parser.
 rm result.csv
 
-python ../cronquot/cronquot.py -s 20161230200000 -e 20170104101111 -d crontab
+python ../cronquot/cronquot.py -s 20161230200000 -e 20170104010000 -d crontab
 
 RESULT=`diff result.csv test/test_result.csv`
